add search template in search.php, search.js uses templete to create new HTML nodes
